fix(dexie): fixed some minor adjustments with dexie
