+++ .env.example
@@ -1,0 +1,6 @@
+NEXT_PUBLIC_API_BASE_URL = http://localhost:3000/api
+NEXT_PUBLIC_FIREBASE_API_KEY = <FIREBASE_API_KEY>
+YOUTUBE_API_KEY = <YOUTUBE_API_KEY>
+GOOGLE_ANALYTICS_PROPERTY_ID = <PROPERTY_ID>
+GOOGLE_CLIENT_EMAIL = <EMAIL_ID>
+GOOGLE_ANALYTICS_PRIVATE_KEY = <PRIVATE_KEY>
